Updated gravatar image handling and updated npm dependencies
