+++ lib/vutuv_web/controllers/user_controller.ex
@@ -33,6 +33,7 @@ defmodule VutuvWeb.UserController do
       {:ok, user} ->
         Log.info(%Log{user: user.id, message: "user created"})
         email = user_params["email"]
+        # maybe move these lines to confirm/new
         user_credential = Accounts.get_user_credential(%{"user_id" => user.id})
         code = Otp.create(user_credential.otp_secret)
         Email.confirm_request(email, code)
