+++ priv/gettext/de/LC_MESSAGES/default.po
@@ -17,11 +17,6 @@ msgid "Already signed up?"
 msgstr "Schon ...?"
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:41
-msgid "Avatar:"
-msgstr "Avatar:"
-
-#, elixir-format
 #: lib/vutuv_web/templates/email_address/edit.html.eex:5
 #: lib/vutuv_web/templates/email_address/new.html.eex:5
 #: lib/vutuv_web/templates/email_address/show.html.eex:38
@@ -60,7 +55,7 @@ msgid "Gender:"
 msgstr "Geschlecht"
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:46
+#: lib/vutuv_web/templates/user/show.html.eex:41
 msgid "Headline:"
 msgstr "Headline:"
 
@@ -80,7 +75,7 @@ msgid "Locale:"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:46
+#: lib/vutuv_web/templates/layout/app.html.eex:52
 #: lib/vutuv_web/templates/session/new.html.eex:21
 msgid "Login"
 msgstr "Einloggen"
@@ -91,7 +86,7 @@ msgid "Login here"
 msgstr "Einloggen hier"
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:42
+#: lib/vutuv_web/templates/layout/app.html.eex:48
 msgid "Logout"
 msgstr "Ausloggen"
 
@@ -101,7 +96,7 @@ msgid "New user"
 msgstr "Neue User"
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:51
+#: lib/vutuv_web/templates/user/show.html.eex:46
 msgid "Noindex?:"
 msgstr ""
 
@@ -125,7 +120,6 @@ msgstr "Anzeigen"
 #, elixir-format
 #: lib/vutuv_web/templates/email_address/index.html.eex:6
 #: lib/vutuv_web/templates/user/index.html.eex:6
-#: lib/vutuv_web/templates/user/show.html.eex:1
 msgid "User"
 msgstr "User"
 
@@ -175,12 +169,12 @@ msgid "Sign up"
 msgstr "Anmelden"
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:37
+#: lib/vutuv_web/templates/layout/app.html.eex:43
 msgid "About"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:38
+#: lib/vutuv_web/templates/layout/app.html.eex:44
 msgid "Data privacy"
 msgstr ""
 
@@ -261,18 +255,19 @@ msgid "is public:"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:66
+#: lib/vutuv_web/templates/user/show.html.eex:62
 msgid "Edit email address"
 msgstr "Ändern E-Mail"
 
 #, elixir-format
 #: lib/vutuv_web/templates/user/edit.html.eex:1
-#: lib/vutuv_web/templates/user/show.html.eex:57
+#: lib/vutuv_web/templates/user/show.html.eex:54
 msgid "Edit user"
 msgstr "Ändern"
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:61
+#: lib/vutuv_web/templates/user/show.html.eex:57
+#: lib/vutuv_web/templates/user/show.html.eex:71
 msgid "Email addresses"
 msgstr "Neue E-Mail Adresse"
 
@@ -282,7 +277,7 @@ msgid "Listing users"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:70
+#: lib/vutuv_web/templates/user/show.html.eex:66
 msgid "New email address"
 msgstr "Neue E-Mail Adresse"
 
@@ -306,7 +301,7 @@ msgstr "Sonstiges"
 
 #, elixir-format
 #: lib/vutuv_web/templates/error/404.html.eex:15
-#: lib/vutuv_web/templates/layout/app.html.eex:26
+#: lib/vutuv_web/templates/layout/app.html.eex:32
 msgid "The fast, secure and less annoying social network."
 msgstr ""
 
@@ -439,4 +434,14 @@ msgstr ""
 #, elixir-format
 #: lib/vutuv_web/templates/password_reset/edit.html.eex:19
 msgid "Update password"
+msgstr ""
+
+#, elixir-format, fuzzy
+#: lib/vutuv_web/templates/user/show.html.eex:12
+msgid "None"
+msgstr ""
+
+#, elixir-format
+#: lib/vutuv_web/templates/confirm/new.html.eex:23
+msgid "Resend code"
 msgstr ""
