+++ priv/gettext/default.pot
@@ -16,11 +16,6 @@ msgid "Already signed up?"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:41
-msgid "Avatar:"
-msgstr ""
-
-#, elixir-format
 #: lib/vutuv_web/templates/email_address/edit.html.eex:5
 #: lib/vutuv_web/templates/email_address/new.html.eex:5
 #: lib/vutuv_web/templates/email_address/show.html.eex:38
@@ -59,7 +54,7 @@ msgid "Gender:"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:46
+#: lib/vutuv_web/templates/user/show.html.eex:41
 msgid "Headline:"
 msgstr ""
 
@@ -79,7 +74,7 @@ msgid "Locale:"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:46
+#: lib/vutuv_web/templates/layout/app.html.eex:52
 #: lib/vutuv_web/templates/session/new.html.eex:21
 msgid "Login"
 msgstr ""
@@ -90,7 +85,7 @@ msgid "Login here"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:42
+#: lib/vutuv_web/templates/layout/app.html.eex:48
 msgid "Logout"
 msgstr ""
 
@@ -100,7 +95,7 @@ msgid "New user"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:51
+#: lib/vutuv_web/templates/user/show.html.eex:46
 msgid "Noindex?:"
 msgstr ""
 
@@ -124,7 +119,6 @@ msgstr ""
 #, elixir-format
 #: lib/vutuv_web/templates/email_address/index.html.eex:6
 #: lib/vutuv_web/templates/user/index.html.eex:6
-#: lib/vutuv_web/templates/user/show.html.eex:1
 msgid "User"
 msgstr ""
 
@@ -174,12 +168,12 @@ msgid "Sign up"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:37
+#: lib/vutuv_web/templates/layout/app.html.eex:43
 msgid "About"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/layout/app.html.eex:38
+#: lib/vutuv_web/templates/layout/app.html.eex:44
 msgid "Data privacy"
 msgstr ""
 
@@ -260,18 +254,19 @@ msgid "is public:"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:66
+#: lib/vutuv_web/templates/user/show.html.eex:62
 msgid "Edit email address"
 msgstr ""
 
 #, elixir-format
 #: lib/vutuv_web/templates/user/edit.html.eex:1
-#: lib/vutuv_web/templates/user/show.html.eex:57
+#: lib/vutuv_web/templates/user/show.html.eex:54
 msgid "Edit user"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:61
+#: lib/vutuv_web/templates/user/show.html.eex:57
+#: lib/vutuv_web/templates/user/show.html.eex:71
 msgid "Email addresses"
 msgstr ""
 
@@ -281,7 +276,7 @@ msgid "Listing users"
 msgstr ""
 
 #, elixir-format
-#: lib/vutuv_web/templates/user/show.html.eex:70
+#: lib/vutuv_web/templates/user/show.html.eex:66
 msgid "New email address"
 msgstr ""
 
@@ -305,7 +300,7 @@ msgstr ""
 
 #, elixir-format
 #: lib/vutuv_web/templates/error/404.html.eex:15
-#: lib/vutuv_web/templates/layout/app.html.eex:26
+#: lib/vutuv_web/templates/layout/app.html.eex:32
 msgid "The fast, secure and less annoying social network."
 msgstr ""
 
@@ -438,4 +433,14 @@ msgstr ""
 #, elixir-format
 #: lib/vutuv_web/templates/password_reset/edit.html.eex:19
 msgid "Update password"
+msgstr ""
+
+#, elixir-format
+#: lib/vutuv_web/templates/user/show.html.eex:12
+msgid "None"
+msgstr ""
+
+#, elixir-format
+#: lib/vutuv_web/templates/confirm/new.html.eex:23
+msgid "Resend code"
 msgstr ""
